Fix a crash case that occurs sporadically with large requests.
The root cause is the fact that apache http client first tries a request with no authentication and then tries the request with basic authentication.  Anaimoto only uses basic authentication, so this also will be a performance improvement.  In addition, make the API handle weird server error codes better.  Also, include the new SubmitJob program, which should make future problems easier to reproduce as it allows a job to be specified on the command-line.

Bump up the version to 1.2.2.